Fix autoloading for core traits. Move `eval(...php:boot...)` to `_bootstrap.php`
