+++ src/CRM/CivixBundle/Resources/codeception/tests/_support/CvBoot.php
@@ -1,51 +1,5 @@
 <?php
 
-/**
- * Call the "cv" command.
- *
- * @param string $cmd
- *   The rest of the command to send.
- * @param string $decode
- *   Ex: 'json' or 'phpcode'.
- * @return string
- *   Response output (if the command executed normally).
- * @throws \RuntimeException
- *   If the command terminates abnormally.
- */
-function cv($cmd, $decode = 'json') {
-  return _cv_run("cv $cmd", $decode);
-}
-
-function _cv_run($cmd, $decode = 'json') {
-  $descriptorSpec = array(0 => array("pipe", "r"), 1 => array("pipe", "w"), 2 => STDERR);
-  $oldOutput = getenv('CV_OUTPUT');
-  putenv("CV_OUTPUT=json");
-  $process = proc_open($cmd, $descriptorSpec, $pipes, __DIR__);
-  putenv("CV_OUTPUT=$oldOutput");
-  fclose($pipes[0]);
-  $result = stream_get_contents($pipes[1]);
-  fclose($pipes[1]);
-  if (proc_close($process) !== 0) {
-    throw new RuntimeException("Command failed ($cmd):\n$result");
-  }
-  switch ($decode) {
-    case 'raw':
-      return $result;
-
-    case 'phpcode':
-      // If the last output is /*PHPCODE*/, then we managed to complete execution.
-      if (substr(trim($result), 0, 12) !== "/*BEGINPHP*/" || substr(trim($result), -10) !== "/*ENDPHP*/") {
-        throw new \RuntimeException("Command failed ($cmd):\n$result");
-      }
-      return $result;
-
-    case 'json':
-      return json_decode($result, 1);
-
-    default:
-      throw new RuntimeException("Bad decoder format ($decode)");
-  }
-}
 
 /**
  * @file
@@ -90,7 +44,6 @@ class CvBoot extends \Codeception\Extension {
   private $startUrl = NULL;
 
   public function beforeSuite(\Codeception\Event\SuiteEvent $e) {
-    $this->boot();
   }
 
   public function beforeTest(\Codeception\Event\TestEvent $e) {
@@ -113,20 +66,6 @@ class CvBoot extends \Codeception\Extension {
           'url' => $this->getStartUrl(),
         ));
       }
-    }
-  }
-
-  protected function boot() {
-    static $booted = FALSE;
-    if ($booted) {
-      return;
-    }
-    $booted = TRUE;
-
-    $extConfig = $this->getExtConfig();
-    if (!empty($extConfig['command'])) {
-      //$this->writeln("\n\nBOOT\n\n");
-      eval(_cv_run($extConfig['command'], 'phpcode'));
     }
   }
