+++ src/CRM/CivixBundle/Resources/codeception/tests/acceptance/_bootstrap.php
@@ -1,2 +1,3 @@
 <?php
 // Here you can initialize variables that will be available to your tests
+eval(_cv_run('cv php:boot', 'phpcode'));
